bool solution
Very simple, pushes the processing to the user which might be simpler for their code. Downside is that it requires users understanding that govuk-font includes our typography extras. It also feels a bit weird that users can pick a font size for their own break points but we're giving them the size "from tablet".