Extra param solution
Feels simple to implement however quickly becomes difficult to manage and understand. What is $size in this instance? The mobile/default value? The value at the breakpoint it's being used at?

This and previous solutions ahve also exposed a problem with approach that we want users to pick different values for different breakpoints but the way they interact with our scale to achieve this would be via different scale values, which are themselves maps of breakpoints. We need to be more intentional in targeting scale values AND breakpoints.