size override solution
The user now is picking one of the maps within the breakpoint maps within the scale maps which then overrides a specific breakpoint in the selected map. This makes the most sense to me as a solution as we're exposing the actual values within the scale rather than trying to presume what a user expects based on the values they're sending.

This requires an understanding of the scale to use effectively but this can be solved by us documenting the scale more clearly.

We would need to think about how line height interacts with this, maybe it needs to be a map as well.

Could we simplify this by implementing the philosophy of this solution into previous solutions eg: passing a map to $size?